[EC-428] Moved CreateProviderCommand to namespace Bit.Commercial.Core.Providers
