+++ bitwarden_license/src/Commercial.Core/Providers/CreateProviderCommand.cs
@@ -1,11 +1,11 @@
 ﻿using Bit.Core.Entities.Provider;
 using Bit.Core.Enums.Provider;
 using Bit.Core.Exceptions;
-using Bit.Core.ProviderFeatures.Interfaces;
+using Bit.Core.Providers.Interfaces;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
 
-namespace Bit.Commercial.Core.ProviderFeatures;
+namespace Bit.Commercial.Core.Providers;
 
 public class CreateProviderCommand : ICreateProviderCommand
 {
